Provider: add initial items
